requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
beautifulsoup4==4.12.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
fastapi==0.115.5
frozenlist==1.5.0
google-api-core==2.23.0
google-api-python-client==2.100.0
google-auth==2.23.4
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.2.1
google-cloud-texttospeech==2.13.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.62.3
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.7.1
MarkupSafe==3.0.2
multidict==6.1.0
notion-client==2.0.0
oauthlib==3.2.2
openai==1.54.5
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyparsing==3.2.0
python-dotenv==1.0.1
python-multipart==0.0.17
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
tqdm==4.67.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
yarl==1.17.2